reorganize the code into packages
